pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-service_incident"
version = "0.1.0"
description = "Service Incident environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
inference = [
    "openai>=1.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m service_incident.server.app
server = "service_incident.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["service_incident", "service_incident.server"]
package-dir = { "service_incident" = ".", "service_incident.server" = "server" }